Initial commit.
Ruby packaging works. The Rakefile creats a vendor/ directory based on
the Jarfile, which is included in the gem. bin/crash sets up the
environment for running with a cluster and standalone. There is no
distinction bewteen cluster and standalone gems, so the gem is huge.

Also, because there is no python version, the python-standalone jar is
currently included.

crash.sh does the same thing as crash without ruby. It could be used to
package either version (or both) in a parcel. (It expects to be in bin/)